# backend/contest_training_system/docs/api.yml
openapi: 3.0.0
info:
  title: 竞赛集训系统 API
  description: '根据提供的Markdown文档生成的完整版竞赛集训系统RESTful API'
  version: '1.0.0'
servers:
  - url: https://api.example.com
    description: '生产服务器'
...
paths:
...
  /register:
    post:
      tags:
        - '1. 用户管理'
      summary: '1.1 注册用户'
      description: '用户注册，创建新账号'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - phoneNumber
                - nickname
                - studentId
                - password
              properties:
                email:
                  type: string
                  description: '邮箱'
                  example: 'test@example.com'
                phoneNumber:
                  type: string
                  description: '手机号'
                  example: '13800000000'
                nickname:
                  type: string
                  description: '昵称'
                  example: '张三'
                studentId:
                  type: string
                  description: '学号'
                  example: '20250001'
                password:
                  type: string
                  description: '密码'
                  example: '123456'
      responses:
        '200':
          description: '注册成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    description: '响应码，1成功，0失败'
                    example: 1
                  msg:
                    type: string
                    description: '提示信息'
                    example: '注册成功'
                  data:
                    type: object
                    properties:
                      userId:
                        type: integer
                        example: 1001
                      email:
                        type: string
                        example: 'test@example.com'
                      nickname:
                        type: string
                        example: '张三'
  /login:
    post:
      tags:
        - '1. 用户管理'
      summary: '1.2 用户登录'
      description: '用户登录，获取token'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - emailOrPhone
                - password
              properties:
                emailOrPhone:
                  type: string
                  description: '邮箱或手机号'
                  example: 'test@example.com'
                password:
                  type: string
                  description: '密码'
                  example: '123456'
      responses:
        '200':
          description: '登录成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: '登录成功'
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                        example: 'xxxxxx'
  /user/deactivate:
    post:
      tags:
        - '1. 用户管理'
      summary: '1.3 注销用户'
      description: '注销当前用户'
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: '用户token'
                  example: 'xxxxxx'
      responses:
        '200':
          description: '注销成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: '注销成功'
                  data:
                    type: 'object'
                    nullable: true
  /user/update:
    put:
      tags:
        - '1. 用户管理'
      summary: '1.4 更新用户信息'
      description: '更新用户基本信息'
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                  description: '新昵称'
                  example: '新昵称'
                phoneNumber:
                  type: string
                  description: '新手机号'
                  example: '13900000000'
                password:
                  type: string
                  description: '新密码'
                  example: 'newpassword'
      responses:
        '200':
          description: '更新成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: '更新成功'
                  data:
                    type: object
                    properties:
                      userId:
                        type: integer
                        example: 1001
                      nickname:
                        type: string
                        example: '新昵称'
                      phoneNumber:
                        type: string
                        example: '13900000000'
  /user/profile:
    get:
      tags:
        - '1. 用户管理'
      summary: '1.5 获取用户信息'
      description: '获取当前用户详细信息'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: '获取成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: 'success'
                  data:
                    type: object
                    properties:
                      userId:
                        type: integer
                        example: 1001
                      email:
                        type: string
                        example: 'test@example.com'
                      phoneNumber:
                        type: string
                        example: '13800000000'
                      nickname:
                        type: string
                        example: '张三'
                      studentId:
                        type: string
                        example: '20250001'
                      role:
                        type: string
                        example: 'STUDENT'
                      createdAt:
                        type: string
                        format: 'date-time'
                        example: '2025-01-01 10:00:00'
  /admin/users:
    get:
      tags:
        - '1. 用户管理'
      summary: '1.6 查询用户列表（管理员）'
      description: '管理员查询用户列表，支持分页、筛选和排序'
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: '页码'
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
          description: '每页数量'
        - name: keyword
          in: query
          schema:
            type: string
          description: '搜索关键字（昵称/邮箱/学号）'
        - name: role
          in: query
          schema:
            type: string
          description: '用户角色筛选'
        - name: isActive
          in: query
          schema:
            type: integer
            enum: [0, 1]
          description: '用户状态筛选（0/1）'
        - name: sortBy
          in: query
          schema:
            type: string
            enum:
              [id_asc, id_desc, nickname_asc, nickname_desc, created_asc, created_desc]
            default: 'id_asc'
          description: '排序字段'
      responses:
        '200':
          description: '查询成功'
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 1
                  msg:
                    type: string
                    example: 'success'
                  data:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 100
                      page:
                        type: integer
                        example: 1
                      pageSize:
                        type: integer
                        example: 20
                      users:
                        type: array
                        items:
                          type: object
                          properties:
                            userId:
                              type: integer
                              example: 1001
                            email:
                              type: string
                              example: 'test@example.com'
                            nickname:
                              type: string
                              example: '张三'
                            studentId:
                              type: string
                              example: '20250001'
                            role:
                              type: string
                              example: 'STUDENT'
                            isActive:
                              type: integer
                              example: 1
                            createdAt:
                              type: string
                              format: date-time
                              example: '2025-01-01 10:00:00'

  # 2. 赛事管理
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '在需要鉴权的接口的请求头中添加 "Authorization: Bearer <YOUR_TOKEN>"'